requirements.txt
flask==3.1.1
python-dotenv==1.1.0
requests==2.32.3
ollama==0.4.8
h2==4.1.0
httpx==0.27.0
gunicorn==21.2.0
gevent==24.2.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic>=2.9.0,<3.0.0 